+++ code/src/cljs/sixsq/nuvla/ui/profile/views.cljs
@@ -1124,8 +1124,11 @@
         loading?      (subscribe [::subs/loading? :customer-info])
         customer-info (subscribe [::subs/customer-info])]
     (fn []
-      (let [{:keys [street-address city country postal-code]} (:address @customer-info)
-            {:keys [fullname balance currency] :or {balance 0}} @customer-info
+      (let [{:keys [fullname balance address
+                    currency email] :or {balance 0}} @customer-info
+            {:keys [street-address city country
+                    postal-code]} address
+
             balance (* balance -1)]
         [ui/Segment {:padded  true
                      :color   "grey"
@@ -1140,6 +1143,10 @@
            [ui/TableRow
             [ui/TableCell [:b (@tr [:street-address])]]
             [ui/TableCell street-address [:br] postal-code " - " city [:br] country]]
+           (when email
+             [ui/TableRow
+              [ui/TableCell [:b "Email"]]
+              [ui/TableCell email]])
            [ui/TableRow {:negative (neg? balance)
                          :positive (pos? balance)}
             [ui/TableCell {:width 5} [:b (str/capitalize (@tr [:balance]))]]